implementations p.2 - shallow vocabcache - tokenization test for spark w2v
Former-commit-id: b744c60394e30eb90e00802c771e8470b9438560 [formerly af60b77b09392da7f1648f027f75bd629f302409]
Former-commit-id: f477a2f8bd22ec3c9b7e54e516773a9d84fdb13b

+++ deeplearning4j-nlp-parent/deeplearning4j-nlp/src/main/java/org/deeplearning4j/models/embeddings/loader/WordVectorSerializer.java
@@ -1029,7 +1029,7 @@ public class WordVectorSerializer {
                 codes.add(Byte.parseByte(split[i]));
             }
             word.setCodes(codes);
-            word.setCodeLength(codes.size());
+            word.setCodeLength((short) codes.size());
         }
         reader.close();
